Deployer: Correct edit parameter fallback
If the old name and new name are given, but we are not POSTing a request
then use the old name to find the deployment again. Use case scenario is
when a user submits a form but ends up being logged out (cookie expiry),
previously the deployment would not be found after login, whereas now we
at least provide the unchanged state again after the user is redirected.